MSI-877: Decrease coupling between objects for Source/Save controller.
